api/pyproject.toml
[project]
name = "fitness"
version = "0.1.0"
description = "Analysis of my fitness data"
readme = "README.md"
requires-python = ">=3.12.0"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "httpx>=0.28.1",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "psycopg[binary]>=3.2.0",
    "alembic>=1.14.0",
    "typing_extensions>=4.0.0",
    "PyJWT[crypto]>=2.8.0",
    "cryptography>=42.0.0",
]

[project.scripts]
app = "fitness.app:app"

[dependency-groups]
dev = [
    "coverage>=7.9.2",
    "ipython>=9.4.0",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "ruff>=0.12.3",
    "ty>=0.0.1a16",
    "testcontainers[postgres]>=4.6.0",
    "pytest-asyncio>=1.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
markers = [
    "e2e: end-to-end API + DB workflow tests",
]
addopts = "--cov=fitness --cov-report html"
